SearchDao interface

+++ src/main/java/com/room414/racingbets/dal/concrete/mysql/infrastructure/MySqlDaoHelper.java
@@ -57,11 +57,11 @@ public class MySqlDaoHelper {
         }
     }
 
-    static long getCount(PreparedStatement statement) throws SQLException {
+    static int getCount(PreparedStatement statement) throws SQLException {
         return getResult(statement, MySqlMapHelper::mapCount);
     }
 
-    static long executeUpdate(PreparedStatement statement) throws SQLException {
+    static int executeUpdate(PreparedStatement statement) throws SQLException {
         return statement.executeUpdate();
     }
